conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.5=py310h9128911_0
brotli=1.0.9=hcfcfb64_8
brotli-bin=1.0.9=hcfcfb64_8
brotlipy=0.7.0=py310h8d17308_1005
bzip2=1.0.8=h8ffe710_4
ca-certificates=2023.01.10=haa95532_0
certifi=2022.12.7=py310haa95532_0
cffi=1.15.1=py310h628cb3f_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.3=pyhd8ed1ab_0
contourpy=1.0.7=py310h232114e_0
cpuonly=2.0=0
cryptography=39.0.2=py310h6e82f81_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.6.6=py310h00ffb61_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fonttools=4.39.0=py310h8d17308_0
freetype=2.12.1=h546665d_1
future=0.18.3=pyhd8ed1ab_0
gettext=0.21.1=h5728263_0
glib=2.74.1=h12be248_1
glib-tools=2.74.1=h12be248_1
gpytorch=1.9.1=pyhd8ed1ab_0
gst-plugins-base=1.22.0=h001b923_2
gstreamer=1.22.0=h6b5321d_2
icu=70.1=h0e60522_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.1.0=pyha770c72_0
importlib_metadata=6.1.0=hd8ed1ab_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.22.0=pyh025b116_0
ipython=8.12.0=pyh08f2357_0
jedi=0.18.2=pyhd8ed1ab_0
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=hcfcfb64_3
jupyter_client=8.1.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py310h5588dad_0
kiwisolver=1.4.4=py310h232114e_1
krb5=1.20.1=heb0366b_0
lcms2=2.15=ha5c8aab_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=12_win64_mkl
libbrotlicommon=1.0.9=hcfcfb64_8
libbrotlidec=1.0.9=hcfcfb64_8
libbrotlienc=1.0.9=hcfcfb64_8
libcblas=3.9.0=12_win64_mkl
libclang=15.0.7=default_h77d9078_1
libclang13=15.0.7=default_h77d9078_1
libdeflate=1.17=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
libglib=2.74.1=he8f3873_1
libhwloc=2.9.0=h51c2c0f_0
libiconv=1.17=h8ffe710_0
liblapack=3.9.0=12_win64_mkl
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h19919ed_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.40.0=hcfcfb64_0
libtiff=4.5.0=hf8721a0_2
libuv=1.44.2=h8ffe710_0
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libxml2=2.10.3=hc3477c8_0
libzlib=1.2.13=hcfcfb64_4
linear_operator=0.3.0=pyhd8ed1ab_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.7.1=py310h5588dad_0
matplotlib-base=3.7.1=py310h51140c5_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h0e2418a_729
mkl-service=2.4.0=py310hcf6e17e_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
ninja=1.11.1=h91493d7_0
numexpr=2.8.4=py310hd213c9f_0
numpy=1.24.2=py310hd02465a_0
openjpeg=2.5.0=ha2aaf27_2
openssl=3.1.0=hcfcfb64_3
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.2=py310h4ed8f06_0
parso=0.8.3=pyhd8ed1ab_0
pcre2=10.40=h17e33f8_0
pickleshare=0.7.5=py_1003
pillow=9.4.0=py310hdbb7713_1
pip=23.0.1=pyhd8ed1ab_0
platformdirs=3.1.0=pyhd8ed1ab_0
ply=3.11=py_1
pooch=1.7.0=pyhd8ed1ab_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
psutil=5.9.4=py310h8d17308_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.14.0=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py310h1fd54f2_3
pyqt5-sip=12.11.0=py310h00ffb61_3
pysocks=1.7.1=pyh0701188_6
python=3.10.9=h4de0772_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytorch=1.13.1=py3.10_cpu_0
pytorch-mutex=1.0=cpu
pytz=2022.7=py310haa95532_0
pywin32=304=py310h00ffb61_2
pyyaml=6.0=py310h8d17308_5
pyzmq=25.0.2=py310hcd737a0_0
qt-main=5.15.8=h720456b_6
requests=2.28.2=pyhd8ed1ab_0
scikit-learn=1.2.2=py310had3394f_0
scipy=1.10.1=py310h578b7cb_0
setuptools=67.6.0=pyhd8ed1ab_0
sip=6.7.7=py310h00ffb61_0
six=1.16.0=pyh6c4a22f_0
stack_data=0.6.2=pyhd8ed1ab_0
tbb=2021.8.0=h91493d7_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h8ffe710_0
toml=0.10.2=pyhd8ed1ab_0
tornado=6.2=py310h8d17308_1
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h191b570_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=15.0.0=py310h8d17308_0
urllib3=1.26.14=pyhd8ed1ab_0
vc=14.3=hb6edc58_10
vs2015_runtime=14.34.31931=h4c5c07a_10
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zeromq=4.3.4=h0e60522_1
zipp=3.15.0=pyhd8ed1ab_0
zstd=1.5.2=h12be248_6
